Revert "Merge sponsor mailer into petition mailer"
