+++ app/jobs/sponsor_signed_email_below_threshold_email_job.rb
@@ -1,5 +1,5 @@
 class SponsorSignedEmailBelowThresholdEmailJob < EmailJob
-  self.mailer = PetitionMailer
+  self.mailer = SponsorMailer
   self.email = :sponsor_signed_email_below_threshold
 
   def perform(signature)
