+++ app/jobs/sponsor_signed_email_on_threshold_email_job.rb
@@ -1,5 +1,5 @@
 class SponsorSignedEmailOnThresholdEmailJob < EmailJob
-  self.mailer = PetitionMailer
+  self.mailer = SponsorMailer
   self.email = :sponsor_signed_email_on_threshold
 
   def perform(signature)
