+++ app/views/sponsor_mailer/petition_and_email_confirmation_for_sponsor.text.erb
