+++ app/views/sponsor_mailer/sponsor_signed_email_below_threshold.text.erb
