+++ app/views/sponsor_mailer/sponsor_signed_email_on_threshold.text.erb
